remove(*): remove user browser & change star build
