+++ src/main/Background.ts
@@ -16,7 +16,6 @@ import {
   loadConfig,
 } from "../modules/config/ConfigSupport";
 import { getMainWindow, getMainWindowUATrimmed } from "./Bootstrap";
-import { getUserBrowser, openBrowser } from "./Browser";
 
 const LOGIN_START =
   "https://login.live.com/oauth20_authorize.srf?client_id=00000000402b5328&response_type=code&scope=service%3A%3Auser.auth.xboxlive.com%3A%3AMBI_SSL&redirect_uri=https%3A%2F%2Flogin.live.com%2Foauth20_desktop.srf";
@@ -50,9 +49,6 @@ export function registerBackgroundListeners(): void {
     } catch {}
     try {
       logoutWindow?.destroy();
-    } catch {}
-    try {
-      getUserBrowser()?.destroy();
     } catch {}
     try {
       getMainWindow()?.destroy();
@@ -258,9 +254,6 @@ export function registerBackgroundListeners(): void {
         }
       });
     });
-  });
-  ipcMain.handle("openBrowser", async (_e, node: boolean, proxy: string) => {
-    await openBrowser(node, proxy);
   });
   ipcMain.handle("getMainWindow", () => {
     return getMainWindow()?.webContents.id || 0;
